requirements.txt
selenium
discord.py
youtube_dl
ffmpeg
psutil
profanity-check
cleverbotfree
# sentry-sdk==0.14.1
discord-sentry-reporting
praw
